Add Apollo with example.
