+++ package.json
@@ -14,10 +14,12 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@apollo/client": "^3.6.9",
     "@expo/ngrok": "^4.1.0",
     "expo": "~45.0.0",
     "expo-cli": "^5.4.12",
     "expo-status-bar": "~1.3.0",
+    "graphql": "^15.8.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-native": "0.68.2",
